8/5

+++ MyLibrary/BIT.cpp
@@ -7,17 +7,17 @@ template< typename T >
 struct BinaryIndexedTree {
   vector< T > data;
 
-  BinaryIndexedTree(int sz) {
+  BinaryIndexedTree(ll sz) {
     data.assign(++sz, 0);
   }
 
-  T sum(int k) {
+  T sum(ll k) {
     T ret = 0;
     for(++k; k > 0; k -= k & -k) ret += data[k];
     return (ret);
   }
 
-  void add(int k, T x) {
+  void add(ll k, T x) {
     for(++k; k < data.size(); k += k & -k) data[k] += x;
   }
 };
@@ -26,7 +26,7 @@ int main()
 {
 ll n=6,ans=0;
 //初期化
-BinaryIndexedTree< int > bit(n);
+BinaryIndexedTree< ll > bit(n);
 //転倒数を求める 解10
 ll tab[5]={5,4,3,2,1};
 rep(i,5){
